enable source maps via a development target
For some reason, source maps don't work unless inlined in web
extensions:

    https://parceljs.org/recipes/web-extension/#hmr

This means that development builds are not as helpful as they could
be.  So add a "development" target which is activated via the normal
"npm run watch" and is configured to inline source maps.

The production build is unaffected and still builds without source
maps.

Closes #72.